﻿M9301/M9312 Console Controller
-- Implemented the M93xxController class, which provides an API for driving
an M9301/M9312 console via the serial interface.

-- Implemented unit test code for the  M93xxController which can be compiled
as a stand-alone executable using the -DUNIT_TEST flag.

-- Implemented dummy version of LoadFileMode to test the M93xxController
class running on the console adapter hardware.  The code was tested against
simh running a M9312 ROM image, with the simulated PDP-11/05 serial console
looped back to the console adapter hardware using a USB-to-TTL serial adapter.