﻿Added support for external activity LEDs
-- Implemented support for external Tx and Rx activity LEDs.  These signal when
the SCL port is transmitting or receiving characters.

-- Implemented support for an extern system activity LED.  This serves as both
a power indicator and a signal of activity on the USB host or Aux ports.  The
Pico onboard LED mirrors the state of this LED.

The GPIO numbers assigned to the new external LEDs are provisional until the new
revision 4 PCB layout is finalized.