﻿Virtual paper tape reader unmount change
-- Modified the automatic unmount behavior of the virtual paper taper
reader such that the paper tape is unmounted at the moment the PDP-11
outputs the first character after the tape has been fully read. This
gives the user a chance to see the reader progress bar advance to 100%
before being cleared from the screen.

On effect of this is that, when reading a BASIC program using the OLD
command in PDP-11 Paper Tape BASIC, the progress bar will stay on the
screen at 100% until the user presses the stop key (CTRL+P) to end the
reading process.